+++ modules/robot/src/real-robot/biclops/vpRobotBiclops.cpp
@@ -1289,7 +1289,7 @@ vpRobotBiclops::readPositionFile(const std::string &filename, vpColVector &q)
       std::istringstream ss(line);
       std::string key_;
       ss >> key_;
-      for (size_t i=0; i< vpBiclops::ndof; i++)
+      for (unsigned int i=0; i< vpBiclops::ndof; i++)
         ss >> q[i];
       pos_found = true;
       break;
